Reorganize text into logical chunks
It is easier for users when all information about a concept is presented together. It is also
easier for users to read and understand information when it is broken into logical subsections.
Subsections with clear headings allow users to more easily skim to find the information they need.

This commit moves information related to each command from the original paragraph to
a new section labeled with the command name. It also moves additional context that applies
to more than one of the commands to the top of the article.