Update article context
Users viewing this article are likely relatively new to Git. They may be unsure
of how and when to use push, pull, and fetch.

This commit adds context to the top of this article that helps users understand
how these commands are related and what tasks they will use them for. It removes
a sentence that states that push and pull are described as equivalent to avoid
introducing conjecture into the documentation.