Update fetch description
Git fetch gets all changes from the remote repository, including new branches and tags,
and places them in the user's local repository. This commit updates the text to explain
that fetch retrieves all changes, not just the changes on the user's current branch.

This commit also updates the language so that it does not use the term pull to describe
the results from fetch. That could be confusing for users, as pull is a separate Git command
that works differently from fetch.

This commit also updates the context about how to merge changes so that the documentation
does not assume the name of the user's working local branch.

Finally, this commit makes the description more concise and addresses the user as 'you,'
not 'we.'