Update pull description
Users visiting this documentation are likely new to Git commands and may not
remember exactly what a git fetch and git merge do when reading this answer.
This commit describes both actions involved in a git pull using a bullet
format to make the text easier to read.

This commit also removes the extra sentence about how some users prefer to
use git fetch and git merge separately so that they can review changes, because
this documentation already provides that context as part of the fetch subsection.

Finally, this commit makes the text more concise and addresses the user as 'you,'
rather than 'we'.