Update push description
This commit makes the text more concise and addresses the user as 'you'
while removing mentions of 'our' and 'we,' as the author and the user
are not actually working on this task together.

This commit also tells the user how to push successfully by reminding them that
they must update their local branch before pushing and telling them how to do so.
The phrasing "Before you can push" helps flag for the user that this is something
they must do beforehand, reducing the chance that they try to push, fail, and then
have to come back to the documentation to understand why.

Finally, this commit removes the sentences about how Git checks for a remote tracking
branch. The previous text doesn't tell the user what a remote tracking branch is or how to
set one. A better alternative is to make the text more concise by asserting that the changes
will go to the remote tracking branch and provide an in-text link to documentation explaining
tracking branches and how to set them. This allows this section to focus on defining the push
command, while providing additional information for users who may need it.