Update article title and description
It is easier for users to find information when section titles are task-based. That is,
when titles describe actions that the user wants to do while they are working, rather than
specific commands or features. This commit updates the title of this article to be task-based
and updates the description so that it adds additional context instead of just repeating the title.

This will help users better understand how this information is relevant to their work and when
to use git fetch, pull, and  push.